Added Short Sidebar Workaround
Whenever you have a sidebar with less content to fill it with. The existing styling of the website will extend non-sidebar content to the page width once the sidebar content has concluded. This addition is a workaround within shorter sidebars through using an empty div with an adjustable height value. Set the height value to be as tall as you need for the sidebar to extend to the full page length. Example:  height:300px for a height with a 300 pixel (px) value.